Fix health and damage player

+++ Assets/Scripts/Utils/CurrentEnemy.cs
@@ -9,21 +9,21 @@ public class CurrentEnemy
         public static class EnemyLv1
         {
             public static float health = 10;
-            public static float attack = 2;
+            public static int attack = 2;
             public static float speed = 3;
             public static float attackSpeed = 1;
         }
         public static class EnemyLv2
         {
             public static float health = 20;
-            public static float attack = 3;
+            public static int attack = 3;
             public static float speed = 2;
             public static float attackSpeed = 1;
         }
         public static class EnemyLv3
         {
             public static float health = 30;
-            public static float attack = 4;
+            public static int attack = 4;
             public static float speed = 1;
             public static float attackSpeed = 1;
         }
